File paths need to convert to/from UTF-8 when going through Javascript
Incomplete attempt to fix issue #183. Hadouken no longer crashes, and the
torrent file makes it through libtorrent and Hadouken far enough for  the
"torrent.add" callback to be made for it, but somewhere further in the code
the torrent is being dropped.

duk strings are expected to be UTF-8, but on Windows the file paths are UTF-16.
This change uses boost to translate all paths to/from duk into UTF-8. It also
uses boost's fstream implementation to open files because STL does not support
UTF-16 strings.

This change is incomplete; there are still uses of STL's fstream that will
fail when presented with a path with non-English characters.

+++ js/plugins/autoadd.js
@@ -21,28 +21,33 @@ function checkFiles(folder, files) {
 
         if(pattern.test(file)) {
             var buffer = fs.readBuffer(file);
-            var p      = bt.AddTorrentParams.getDefault();
-            var meta   = {};
-            p.torrent  = new bt.TorrentInfo(buffer);
+            if(typeof buffer != "undefined") {
+                var p      = bt.AddTorrentParams.getDefault();
+                var meta   = {};
+                p.torrent  = new bt.TorrentInfo(buffer);
 
-            if(folder.savePath) {
-                p.savePath = folder.savePath;
+                if(folder.savePath) {
+                    p.savePath = folder.savePath;
+                }
+
+                if(folder.label) {
+                    meta.label = folder.label;
+                }
+
+                if(folder.tags) {
+                    meta.tags = folder.tags;
+                }
+
+                logger.info("(AUTOADD) Adding torrent " + p.torrent.name + " from file " + file);
+
+                p.metadata = meta;
+                session.addTorrent(p);
+                if(!folder.keep) {
+                    fs.deleteFile(file);
+                }
             }
-
-            if(folder.label) {
-                meta.label = folder.label;
-            }
-
-            if(folder.tags) {
-                meta.tags = folder.tags;
-            }
-
-            logger.info("(AUTOADD) Adding torrent " + p.torrent.name + " from file " + file);
-
-            p.metadata = meta;
-            session.addTorrent(p);
-            if(!folder.keep) {
-                fs.deleteFile(file);
+            else {
+                logger.info("(AUTOADD) Could not read contents of file " + file);
             }
         }
     }
